added subgenre table

+++ MenrvaBooks/src/main/kotlin/com/menrva/repositories/SubGenreRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SubGenreRepository : JpaRepository<SubGenre, Int> {
+interface SubGenreRepository : JpaRepository<SubGenre, Long> {
     fun findByName(name: String): List<SubGenre>
 }
